Finish to the day-
Added data model for a User Profile and also for a list of activities.

Fixed the issue with the colour of the tab bar. This was due to the tab bar controller being identified as a view controller.

Attempted to connect the API key to the app however it keeps telling me that access is denied and that the app is not authorised. Tried to fix, maybe find an alternative.